requirements/development.txt
# Django and API
Django==5.0.6
djangorestframework==3.15.1
django-cors-headers==4.3.1
psycopg2-binary==2.9.9
python-decouple==3.8
django-environ==0.11.2

# Telegram Bot
python-telegram-bot==21.3
telegram==0.0.1

# Background Tasks
celery==5.3.4
redis==5.0.1

# Image Processing & QR
Pillow==10.3.0
qrcode==7.4.2
cloudinary==1.36.0

# Google Sheets Integration
google-auth==2.29.0
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
google-api-python-client==2.127.0

# Security & Crypto
cryptography==42.0.5
PyJWT==2.8.0

# Time & Date
pytz==2024.1

# Data Validation
marshmallow==3.21.1

# HTTP & API
requests==2.31.0
httpx==0.27.0

# Development
django-extensions==3.2.3
ipython==8.24.0

# Production
gunicorn==22.0.0
whitenoise==6.6.0

# Monitoring
sentry-sdk==2.1.1